dom/file/test/test_stream_tracking.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    var testBuffer = getRandomBuffer(100000);

    for each (let fileStorage in fileStorages) {
      let request = getFileHandle(fileStorage.key, "test.txt");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      let fileHandle = event.target.result;
      fileHandle.onerror = errorHandler;

      let lockedFile = fileHandle.open("readwrite");

      request = lockedFile.write(testBuffer);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      request = fileHandle.getFile();
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      let file = event.target.result;

      let resultBuffer1;
      let resultBuffer2;

      let reader1 = new FileReader();
      reader1.readAsArrayBuffer(file);
      reader1.onerror = errorHandler;
      reader1.onload = function(event)
      {
        resultBuffer1 = event.target.result;

        let reader = new FileReader();
        try {
          reader.readAsArrayBuffer(file);
          ok(false, "Should have thrown!");
        }
        catch (e) {
          ok(e instanceof DOMException, "Got exception.");
          is(e.name, "LockedFileInactiveError", "Good error.");
          is(e.code, 0, "Good error code.");
        }

        if (resultBuffer2) {
          testGenerator.next();
        }
      }

      let reader2 = new FileReader();
      reader2.readAsArrayBuffer(file);
      reader2.onerror = errorHandler;
      reader2.onload = function(event)
      {
        resultBuffer2 = event.target.result;

        let reader = new FileReader();
        try {
          reader.readAsArrayBuffer(file);
          ok(false, "Should have thrown!");
        }
        catch (e) {
          ok(e instanceof DOMException, "Got exception.");
          is(e.name, "LockedFileInactiveError", "Good error.");
          is(e.code, 0, "Good error code.");
        }

        if (resultBuffer1) {
          testGenerator.next();
        }
      }

      yield undefined;

      ok(compareBuffers(resultBuffer1, testBuffer), "Correct data");
      ok(compareBuffers(resultBuffer2, testBuffer), "Correct data");
    }

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
